[refactor] Structural Integration Work
